Fix random engine implementations.

+++ org.jenetix/src/main/java/org/jenetix/random/MRG2Random.java
@@ -104,7 +104,7 @@ public class MRG2Random extends Random32 {
 			long t = modulus.mod(seed);
 			if (t < 0) t += modulus.VALUE;
 
-			_r1 = (int)t;
+			_r1 = t;
 			_r2 = 1;
 		}
