+++ org.jenetix/src/main/java/org/jenetix/random/MRG5Random.java
@@ -127,7 +127,7 @@ public class MRG5Random  extends Random32 {
 			long t = modulus.mod(seed);
 			if (t < 0) t += modulus.VALUE;
 
-			_r1 = (int)t;
+			_r1 = t;
 			_r2 = 1;
 			_r3 = 1;
 			_r4 = 1;
